+++ CHANGELOG.md
@@ -86,6 +86,8 @@ Main (unreleased)
 
 - Fix `pyroscope.receive_http` so it does not restart server if the server configuration has not changed. (@korniltsev)
 
+- Increase default connection limit in `pyroscope.receive_http` from 100 to 16k. (@korniltsev)
+
 v1.10.1
 -----------------